refactor: qi 관련 test 클래스 패키지 이동
