+++ back/src/test/java/com/qmate/domain/quetioninstance/service/QuestionInstanceServiceGetDetailTest.java
@@ -1,4 +1,4 @@
-package com.qmate.questioninstance;
+package com.qmate.domain.quetioninstance.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.eq;
